refactor: restructure native codes folder
