+++ android/src/main/java/com/rnadmob/admob/RNAdMobPackage.java
@@ -6,10 +6,11 @@ import com.facebook.react.ReactPackage;
 import com.facebook.react.bridge.NativeModule;
 import com.facebook.react.bridge.ReactApplicationContext;
 import com.facebook.react.uimanager.ViewManager;
-import com.rnadmob.admob.ads.RNAdMobAppOpenAdModule;
-import com.rnadmob.admob.ads.RNAdMobInterstitialAdModule;
-import com.rnadmob.admob.ads.RNAdMobRewardedAdModule;
-import com.rnadmob.admob.ads.RNAdMobRewardedInterstitialAdModule;
+import com.rnadmob.admob.ads.fullscreen.RNAdMobAppOpenAdModule;
+import com.rnadmob.admob.ads.banner.RNAdMobBannerViewManager;
+import com.rnadmob.admob.ads.fullscreen.RNAdMobInterstitialAdModule;
+import com.rnadmob.admob.ads.fullscreen.RNAdMobRewardedAdModule;
+import com.rnadmob.admob.ads.fullscreen.RNAdMobRewardedInterstitialAdModule;
 
 import java.util.Arrays;
 import java.util.List;
